Fixes an issue with the user page looking bad on small screens.
The .pull-right class caused the .col-md-3 to keep its width the size of
the image within, causing it to look really really bad on small (mobile)
screens. I've replaced the .pull-right with just a basic .col-md-3 but
used the "push" and "pull" classes to put the box on top on small screens
but on the right on large screens.